chore(deps): remove husky
This commit removes the `husky` dependency from both the frontend and backend projects.
It simplifies the project setup and reduces the number of dependencies.
The removal of `husky` does not impact the functionality of the application as it is not currently being used.
This change helps to maintain a clean and efficient project structure.