refactor(manage-rules): integrate dnd-kit-sortable-tree for requirements
Integrate dnd-kit-sortable-tree in the manage-rules feature, replacing previous
implementations with the beautiful-dnd library. This change simplifies the management
and nesting of rules by utilizing the dnd-kit core functionalities.

Removed unused components such as the SortableTreeComponent from the test-component
folder and consolidated styles in globals.css to prepare for the new tree viewimplementation. Also, updated type definitions to include UniqueIdentifier from
@dnd-kit/core to ensure type safety.