feat(ui): add new component for tree structure in manage rules
This commit introduces a new component `NewTree.tsx` in the `manage-rules`
directory. This component is responsible for rendering a tree-like
structure of requirements. It leverages the `he-tree-react` library to
achieve the tree functionality. The component allows for expansion and
contraction of tree nodes, as well as visual indicators for connector
nodes and the ability to drag and drop items.