refactor(manage-rules): BUG FIX. Use e.preventDefault() to stop automatically data update
Change the prop name 'onChange' to 'onUpdate' in NestedRequirementsList component
for better clarity on its functionality. This is part of the ongoing effort to
standardize event handler prop names across components.

BREAKING CHANGE: This commit changes the prop name 'onChange' to 'onUpdate' inNestedRequirementsList, which may affect any code using this component with the
previous prop name.